+++ backend/app/Http/Controllers/UserController.php
@@ -3,37 +3,76 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
-    //
-
- public function index()
+    public function index()
     {
-        return User::all();
+        $users = User::all();
+        return response()->json($users, 200);
     }
 
     public function show($id)
     {
-        return User::findOrFail($id);
+        $user = User::find($id);
+        if ($user) {
+            return response()->json($user, 200);
+        } else {
+            return response()->json(['message' => 'User not found'], 404);
+        }
     }
 
     public function store(Request $request)
     {
-        return User::create($request->all());
+        $validatedData = $request->validate([
+            'name' => 'required|string',
+            'email' => 'required|email|unique:users,email',
+            'password' => 'required|string',
+            'user_type' => ['required', Rule::in(['passenger', 'manager', 'headquarter'])],
+            'coin_amount' => 'required|integer|min:0',
+        ]);
+
+        $validatedData['password'] = Hash::make($validatedData['password']);
+
+        $user = User::create($validatedData);
+
+        return response()->json($user, 201);
     }
 
     public function update(Request $request, $id)
     {
-        $user = User::findOrFail($id);
-        $user->update($request->all());
-        return $user;
+        $user = User::find($id);
+        if ($user) {
+            $validatedData = $request->validate([
+                'name' => 'string',
+                'email' => ['email', Rule::unique('users')->ignore($user->id)],
+                'password' => 'string',
+                'user_type' => [Rule::in(['passenger', 'manager', 'headquarter'])],
+                'coin_amount' => 'integer|min:0',
+            ]);
+
+            if (isset($validatedData['password'])) {
+                $validatedData['password'] = Hash::make($validatedData['password']);
+            }
+
+            $user->update($validatedData);
+            return response()->json($user, 200);
+        } else {
+            return response()->json(['message' => 'User not found'], 404);
+        }
     }
 
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
-        return 204;
+        $user = User::find($id);
+        if ($user) {
+            $user->delete();
+            return response()->json(['message' => 'User deleted'], 200);
+        } else {
+            return response()->json(['message' => 'User not found'], 404);
+        }
     }
 }
